new class attributes
