+++ tests/test_utils.py
@@ -10,25 +10,28 @@ import pytest
 import requests
 import web3
 
-from telliot.utils.eth_utils import RPCEndpoint, Contract
+from telliot.utils.rpc_endpoint import RPCEndpoint, Contract
 
 load_dotenv() #we will replace this with loading from config
 
 with open("abi.json") as f:
     abi = f.read()
 
+network = "mainnet"
+provider = "pokt"
+
 def test_rpc_endpoint():
     '''RPCEndpoint connects to the blockchain'''
     url = os.getenv("NODE_URL")
-    endpt = RPCEndpoint(url)
+    endpt = RPCEndpoint(network=network, provider=provider, url=url)
     endpt.connect()
-    assert endpt.web3.eth.blockNumber > 1
+    assert endpt.web3.eth.block_number > 1
     
 
 def test_very_bad_rpc_url():
     '''an invalid url will raise an exception in RPCEndpoint'''
     url = "this is not a valid rpc url"
-    endpt = RPCEndpoint(url)
+    endpt = RPCEndpoint(network, url)
     endpt.connect()
     #expect bad url error from requests library
     with pytest.raises(requests.exceptions.MissingSchema) as e_info:
@@ -37,7 +40,7 @@ def test_very_bad_rpc_url():
 def test_incomplete_rpc_url():
     '''an incomplete url will raise an exception in RPCEndpoint'''
     url = "https://eth-rinkeby.gateway.pokt.network/v1/lb/"
-    endpt = RPCEndpoint(url)
+    endpt = RPCEndpoint(network, url)
     endpt.connect()
     #expect bad url error from requests library
     with pytest.raises(requests.exceptions.HTTPError) as e_info:
@@ -58,9 +61,12 @@ def test_load_from_config():
 def connect_to_contract(address):
     '''Helper function for connecting to a contract at an address'''
     url = os.getenv("NODE_URL")
-    endpt = RPCEndpoint(url)
+    endpt = RPCEndpoint(network, url)
     endpt.connect()
 
     c = Contract(endpt, address, abi)
     c.connect()
     return c
+
+def connect_to_network():
+    pass
